pyproject.toml
[tool.poetry]
name = "todo-tasks"
version = "0.1.1"
description = ""
authors = ["Milan Suk <Milansuk@email.cz>"]
packages = [ { include = "src" } ]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
pydantic = "^1.7.2"
rich = "^9.1.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
mypy = "^0.790"
black = "^20.8b1"
isort = "^5.6.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
tasks = 'src.entrypoint:run'

[tool.black]
line-length = 120
